Made progress on ControlModuleDetailsActivity
The application is currently in debug mode to bypass opening a bluetooth
socket when selecting a bluetooth device in the initial menu.

*BUG FIXES*
Fixed a major bug that was caused by not properly unregistering the
intent receiver that was registered in the base station connect
activity. The application was getting overrun with intents from the
Bluetooth Scan activity. The BaseStationConnectActivity was reworked so
that the intent filter is unregistered whenever it is not in use. This
has fixed the problem of not being able to upload the newest version of
the application to the Android device.