Made changes to enhance robustness of WHCS application.
The application was crashing if the base station lost power once the
application had already connected and the application tried to send data
over bluetooth. The PipelineErrorHandler interface was added and
implemented by WHCSActivity so that if there is an error, it is passed
from the CommadIssuer to the activity where the activity can cleanup and
end itself.

Fixed a bug that would happen if the application successfully created a
socket on the HC-05 but the base station did not respond that it was in
fact the base station. The application would go into a bad state where
it was occupying the HC-05's socket but was not going into the
ControlModuleListActivity. Now if the AsynchInitialization times out it
stops the issuer and listener.

Added a feature so that if a base station was previously used, the
application searches for that base station on start up. If that base
station can not be found this does not hurt the application.