Added Database and GUI structures for Control Module Groups
The database stores control module groups through a table called
control_module_groupings. This table stores pairs of control module
I.D.s and group numbers.
The ControlModuleDetailsActivity now checks what groups a control module
is in when it is shown in the activity. The appropriate checkboxes
reflect the groups the control module is present in. The user can check
and uncheck control modules to put the control modules in different
groups. Clicking the save button will save these changes in the
database.

Commands to the base station to turn off toggleable control modules now
take action based on the response the base station gives. If the base
station says the command failed then the application does not let the
toggleable switch move to the toggled state. This ensures the
application is always in synch with the system.

*need to do*
The control module group buttons present in the
ControlModuleListActivity do not yet have onClickListeners or functions
that behave in the way that the button press should. Implementing the
functionality of these buttons would be a good next step for the
application.